NN Stuff/requirements.txt.txt
absl-py==0.7.1
astor==0.8.0
atom==0.4.3
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
certifi==2019.6.16
colorama==0.4.1
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
entrypoints==0.3
future==0.17.1
gast==0.2.2
google-pasta==0.1.7
grpcio==1.22.0
h5py==2.9.0
ipykernel==5.1.1
ipython==7.6.1
ipython-genutils==0.2.0
ipywidgets==7.5.0
jedi==0.13.3
Jinja2==2.10.1
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Mako==1.0.10
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.0
mistune==0.8.4
mkl-fft==1.0.12
mkl-random==1.0.2
mkl-service==2.0.2
mock==3.0.5
nbconvert==5.5.0
nbformat==4.4.0
notebook==6.0.0
numpy==1.16.4
olefile==0.46
opt-einsum==2.3.2
pandocfilters==1.4.2
parso==0.5.0
pickleshare==0.7.5
Pillow==6.1.0
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.9.0
Pygments==2.4.2
pygpu==0.7.6
pyparsing==2.4.0
pyreadline==2.1
pyrsistent==0.14.11
python-dateutil==2.8.0
pytz==2019.1
pywinpty==0.5.5
PyYAML==5.1.1
pyzmq==18.0.0
qtconsole==4.5.1
scipy==1.2.1
Send2Trash==1.5.0
six==1.12.0
tb-nightly==1.15.0a20190717
tensorboard==1.12.2
tensorflow==1.12.0
termcolor==1.1.0
terminado==0.8.2
Theano==1.0.3+2.g3e47d39ac.dirty
tornado==6.0.3
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
widgetsnbextension==3.5.0
wincertstore==0.2
wrapt==1.11.2
wxPython==4.0.4